feat(unmarshal): add JSON unmarshalling functions with error handling
